+++ packages/flterm/lib/src/foundation.dart
@@ -7,4 +7,5 @@ export 'foundation/input_types.dart';
 export 'foundation/platform_map.dart';
 export 'foundation/terminal_config.dart';
 export 'foundation/terminal_gesture_settings.dart';
+export 'foundation/terminal_keyboard_event.dart';
 export 'foundation/terminal_theme.dart';